@@ -1,11 +1,17 @@
 import React from 'react'
-import ReactDOM from 'react-dom'
-import './index.css'
-import App from './App'
+import { createRoot } from 'react-dom/client'
 
-ReactDOM.render(
+import ToastContextProvider from './context/ToastContextProvider'
+import App from './App'
+import './index.css'
+
+const container = document.getElementById('root')
+const root = createRoot(container)
+
+root.render(
   <React.StrictMode>
-    <App />
+    <ToastContextProvider>
+      <App />
+    </ToastContextProvider>
   </React.StrictMode>,
-  document.getElementById('root'),
 )
